requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
Django==3.2.7
gunicorn==20.1.0
numpy==1.21.1
pandas==1.3.1
pdfkit==0.6.1
Pillow==8.4.0
pytz==2021.1
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.2
sklearn==0.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7